Fix PyThreadState_Get errors
Turns out the best way to obtain compatibility across Python
distributions, it seems, is to not link libpython*.dylib at all! If we
instruct the linker not to worry about undefined symbols, it turns out
the symbols will all come from the running Python process at import
time, and it works across a variety of Python distributions.