Remove the confusing wrapper classes
There is a lot of confusion in various parts of the code around whether
we should be working with raw protobuf-generated data structures,
our own higher level wrappers, or both. Let's standardize on just
the generated protobuf structures, with utility functions for all the
wrapper functionality that can operate directly on protobuf structs.

Leaving TreeEnsemble wrapper class as it uses its state more
extensively. All the others are refactored into utility
functions that operate on protobuf-generated classes.

This commit also:
* Updates .gitignore to ignore ninja build files
* Updates CMakeLists to correctly detect the OS SDK and deployment target
  (resolves build issue on 10.14 from using MLFeatureTypeSequence)