Add a script to sync PRs to Gitlab for CI
The current CI setup only seems to kick off builds for commits that have
been pushed to either a branch on apple/coremltools on GitHub, or a
branch on zach_nation/coremltools on Gitlab. This script will enumerate
all of the open PRs on GitHub apple/coremltools, and push their commits
to a corresponding branch on zach_nation/coremltools.

Requires a valid GitHub access token to be set as an environment
variable. This is to use the GitHub API, but it doesn't need any access
beyond PRs on public repos.

Requires push access to zach_nation/coremltools. Only zach_nation can
give you this.