Resolve AttributeError in cast for Float Types in Non-Tensor Contexts (#2087)
* Improve type handling for cast operation in elementwise_unary.py

This commit refactors the `cast` operation in `elementwise_unary.py` within the coremltools converter. The modification ensures that `input_var.val` is treated appropriately based on its type and the tensor status of `input_var.sym_type`. Specifically, it introduces a conditional check to handle cases where `input_var.val` is a float or `input_var.sym_type` is a tensor, converting `input_var.val` to a NumPy array before casting its type.

* Add test for cast operation with non-tensor float input

This commit introduces a test case in `test_elementwise_unary.py` for the `cast` operation when handling a non-tensor float input. The test specifically targets a scenario where `input_var.val` is a float and `input_var.sym_type` is None, which simulates a non-tensor type.

A custom `MockInputVar` class is used to create an input variable with the specified attributes. The test calls `elementwise_unary.cast.get_cast_value` with this mock input and checks if the output matches the expected NumPy array.

This test ensures that the cast operation correctly handles float values when the symbolic type is not a tensor, guarding against the `AttributeError` that would occur without the appropriate handling in the `get_cast_value` method.

* Utilize string_to_nptype function directly for the cast method

Simplify the cast method implementation for better maintainability and
code readability by utilizing string_to_nptype function directly.

* Enhance type casting to support diverse input formats

This commit introduces a more flexible type casting approach to accommodate inputs with and without the `astype` method. Previously, type casting was performed directly by calling a NumPy type constructor on `input_var.val`. While straightforward, this method did not account for inputs that are already numpy arrays or array-like objects supporting the `astype` method, potentially leading to inefficiencies or errors with certain input types.

The updated implementation first checks for the presence of the `astype` method on `input_var.val`. If available, it utilizes this method to cast the type, ensuring optimal compatibility and performance for numpy arrays and array-like inputs. For other types of inputs that do not support `astype`, the original method of directly invoking the NumPy type constructor is used.

This change enhances the robustness of type casting in our codebase, ensuring broader compatibility with varying input types while maintaining the simplicity and efficiency of the casting operation.

Key Changes:
- Use `astype` for inputs supporting this method, optimizing for numpy and array-like objects.
- Fall back to the direct NumPy type constructor for other inputs, maintaining general compatibility.

* Parametrize input and add expected output

Modify test_cast to use parametrize and handle various input types,
improving test coverage. This change helps ensure
comprehensive testing for different data types.