Avoid fp32 cast for Torch div operator
The `div` Torch op was always casting both operands to fp32, even if both operands are of type fp16. This cast should get removed by the `"common::add_fp16_cast"` optimization pass. However, it causes issues during the PyTorch conversion, for example let's say we have a forward method like this:

```python
class Foo:
    def __init__(self):
        super().__init__()
        self.proj = torch.nn.Linear(16, 1)
    def forward(self, x, y): # both fp16 tensors, shape [1, 16]
        r = x / y # r is now fp32
        return self.proj(r) # Problem
```

Now if we have moved the model (and it's parameters) to fp16 with eg. `m = Foo().to(torch.float16)`, we get an error at conversion time:

> In op, of type linear, named linear_0, the named input `bias` must have the same data type as the named input `x`. However, bias has dtype fp16 whereas x has dtype fp32.

This is because the result of the `div` operation stays fp32, and this doesn't match the resulting type of the PyTorch expression.