Clarify joint_compression parameter, fix style
The description of the parameter `joint_compression` in the
`palettize_weights`, `prune_weights`, and `linear_quantize_weights`
functions was difficult to decipher. This commit re-words it.

This commit also makes small changes to the description of
palettize_weights to help it conform to Apple Style (changing "4" to
"four", for example).