Review of new Core ML Tools public API docstrings, part 2 (#2326)
* Clarify joint_compression parameter, fix style

The description of the parameter `joint_compression` in the
`palettize_weights`, `prune_weights`, and `linear_quantize_weights`
functions was difficult to decipher. This commit re-words it.

This commit also makes small changes to the description of
palettize_weights to help it conform to Apple Style (changing "4" to
"four", for example).

* Clarify linear_quantize_activations description

The description of the linear_quantize_activations function was
difficult to parse, so I modified the wording to more clearly explain
the steps the function walks through to return an activation
quantized MLModel instance.

* Fix typos & formatting for 6 API docstrings

This commit fixes small typos and formatting issues, across six API
docstrings:
1. ModulePostTrainingQuantizerConfig
2. PostTrainingQuantizer
3. PostTrainingQuantizerConfig
4. LayerwiseCompressorConfig
5. LayerwiseCompressor
6. ModuleGPTQConfig

It also fixes similar typos found in other API docstrings that live in
the same file as the targeted APIs, and it expands acronyms that may
not be obvious to the user.

* Update LayerwiseCompressorConfig docstring

Tweak wording for `layers` parameter of the `LayerwiseCompressorConfig`
to make it clear to the reader what the acceptable formats of list
items in the `layers` parameter are.